Version 0.3.0 -- Flying Edward (to safety)
This release adds support for ArchLinux, thanks to the precious help of
@korakinos, and @jonbrett. (See pull #8) It does not bring new functionality
other than being ready-to-use on non-Debian systems.

*

It occurs at an interesting moment in history, where a whistleblower name Edward
Snowden is chased by the U.S.A. for leaking secrets about a global surveillance
system operated by the NSA: PRISM. Today, in order to prevent Snowden from
seeking asylum, the U.S.A. and its allies from Europe (Portugal, France, Italy,
and Austria) are abusing the rules of diplomacy. They're threatening the life of
President Evo Morales of Bolivia, by refusing their aerial space to his
presidential plane on its way back from Russia, where Snowden found a temporary
refuge, on the suspicion that the whistleblower might be on board.

That is unprecedented, and as in the reclusion of Julian Assange, demonstrates
once again the unilaterality of the "democratic rule", where a
pseudo-democratic-superpower-gone-out-of-control respects or violates diplomacy,
human rights, or the Constitution as it sees fit.

    The best weapon of a dictatorship is secrecy, but the best
    weapon of a democracy should be the weapon of openness.
    -- Niels Bohr